blackbasta/requirements.txt
pycryptodomex==3.15.0
rsa==4.9
